# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GESTAVICOLE SaaS API - Gestion avicole multi-tenant
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales in tenant
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleListResponse"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleRequest"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /stock:
    get:
      operationId: listStock
      tags: [stock]
      summary: List stock items in tenant
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of stock items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockListResponse"
    post:
      operationId: createStockItem
      tags: [stock]
      summary: Create a stock item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockRequest"
      responses:
        "201":
          description: Stock item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"
...
  /sales/{saleId}:
    put:
      operationId: updateSale
      tags: [sales]
      summary: Update a sale
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleRequest"
      responses:
        "200":
          description: Sale updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete a sale
      security:
        - bearerAuth: []
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sale deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /stock/{stockId}:
    put:
      operationId: updateStockItem
      tags: [stock]
      summary: Update a stock item
      security:
        - bearerAuth: []
      parameters:
        - name: stockId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockRequest"
      responses:
        "200":
          description: Stock item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"
    delete:
      operationId: deleteStockItem
      tags: [stock]
      summary: Delete a stock item
      security:
        - bearerAuth: []
      parameters:
        - name: stockId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Sale:
      type: object
      properties:
        id:
          type: string
        batchId:
          type: string
          nullable: true
        batchName:
          type: string
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        totalAmount:
          type: number
          format: float
        buyerName:
          type: string
        saleDate:
          type: string
          format: date
        type:
          type: string
        notes:
          type: string
          nullable: true
        tenantId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - quantity
        - unitPrice
        - totalAmount
        - buyerName
        - saleDate
        - type
        - tenantId
        - createdAt
    CreateSaleRequest:
      type: object
      properties:
        batchId:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        buyerName:
          type: string
        saleDate:
          type: string
          format: date
        type:
          type: string
        notes:
          type: string
      required:
        - quantity
        - unitPrice
        - buyerName
        - saleDate
        - type
    SaleListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalAmount:
          type: number
          format: float
      required:
        - data
        - total
        - page
        - limit
        - totalAmount
    Stock:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        quantity:
          type: number
          format: float
        unit:
          type: string
        minimumLevel:
          type: number
          format: float
        currentValue:
          type: number
          format: float
        tenantId:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - quantity
        - unit
        - minimumLevel
        - currentValue
        - tenantId
        - updatedAt
    CreateStockRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        quantity:
          type: number
          format: float
        unit:
          type: string
        minimumLevel:
          type: number
          format: float
        unitPrice:
          type: number
          format: float
      required:
        - name
        - category
        - quantity
        - unit
        - minimumLevel
        - unitPrice
    StockListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Stock"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        lowStockCount:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - lowStockCount
...
    DeleteResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...